Other: Extract `focusDropdownContentsOnArrows()` to own file.
